move to FrameworkBundle (#9)
* move to FrameworkBundle

* review

* test
