+++ composer.json
@@ -73,7 +73,7 @@
         "symfony/error-handler": "self.version",
         "symfony/event-dispatcher": "self.version",
         "symfony/expression-language": "self.version",
-        "symfony/feature-toggle": "self.version",
+        "symfony/feature-flags": "self.version",
         "symfony/filesystem": "self.version",
         "symfony/finder": "self.version",
         "symfony/form": "self.version",
