--- src/Symfony/Bundle/FeatureFlagsBundle/.gitattributes
@@ -1,4 +1,0 @@
-/Tests export-ignore
-/phpunit.xml.dist export-ignore
-/.gitattributes export-ignore
-/.gitignore export-ignore
