--- src/Symfony/Bundle/FeatureFlagsBundle/.gitignore
@@ -1,3 +1,0 @@
-vendor/
-composer.lock
-phpunit.xml
