--- src/Symfony/Bundle/FeatureFlagsBundle/FeatureFlagsBundle.php
@@ -1,24 +1,0 @@
-<?php
-
-/*
- * This file is part of the Symfony package.
- *
- * (c) Fabien Potencier <fabien@symfony.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Symfony\Bundle\FeatureFlagsBundle;
-
-use Symfony\Bundle\FeatureFlagsBundle\DependencyInjection\CompilerPass\DebugPass;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-final class FeatureFlagsBundle extends Bundle
-{
-    public function build(ContainerBuilder $container): void
-    {
-        $container->addCompilerPass(new DebugPass());
-    }
-}
