--- src/Symfony/Bundle/FeatureFlagsBundle/README.md
@@ -1,18 +1,0 @@
-FeatureFlagsBundle
-==================
-
-FeatureFlagsBundle provides a tight integration of the Symfony FeatureFlags
-component into the Symfony full-stack framework.
-
-**This Bundle is experimental**.
-[Experimental features](https://symfony.com/doc/current/contributing/code/experimental.html)
-are not covered by Symfony's
-[Backward Compatibility Promise](https://symfony.com/doc/current/contributing/code/bc.html).
-
-Resources
----------
-
-* [Contributing](https://symfony.com/doc/current/contributing/index.html)
-* [Report issues](https://github.com/symfony/symfony/issues) and
-  [send Pull Requests](https://github.com/symfony/symfony/pulls)
-  in the [main Symfony repository](https://github.com/symfony/symfony)
