--- src/Symfony/Bundle/FeatureFlagsBundle/composer.json
@@ -1,37 +1,0 @@
-{
-    "name": "symfony/feature-flags-bundle",
-    "type": "symfony-bundle",
-    "description": "Provides a tight integration of the Symfony FeatureFlags component into the Symfony full-stack framework",
-    "keywords": [],
-    "homepage": "https://symfony.com",
-    "license": "MIT",
-    "authors": [
-        {
-            "name": "Fabien Potencier",
-            "email": "fabien@symfony.com"
-        },
-        {
-            "name": "Symfony Community",
-            "homepage": "https://symfony.com/contributors"
-        }
-    ],
-    "require": {
-        "php": ">=8.2",
-        "symfony/contracts": "^3.3",
-        "symfony/config": "^6.3",
-        "symfony/dependency-injection": "^6.3",
-        "symfony/http-kernel": "^6.3"
-    },
-    "require-dev": {
-        "symfony/expression-language": "^6.3",
-        "symfony/http-foundation": "^6.3",
-        "symfony/twig-bridge": "^6.3"
-    },
-    "autoload": {
-        "psr-4": { "Symfony\\Bundle\\FeatureFlagsBundle\\": "" },
-        "exclude-from-classmap": [
-            "/Tests/"
-        ]
-    },
-    "minimum-stability": "dev"
-}
