+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Compiler/FeatureFlagsDebugPass.php
@@ -9,29 +9,21 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Bundle\FeatureFlagsBundle\DependencyInjection\CompilerPass;
+namespace Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler;
 
-use Symfony\Bundle\FeatureFlagsBundle\Debug\TraceableFeatureChecker;
-use Symfony\Bundle\FeatureFlagsBundle\Debug\TraceableStrategy;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Reference;
+use Symfony\Component\FeatureFlags\Debug\TraceableFeatureChecker;
+use Symfony\Component\FeatureFlags\Debug\TraceableStrategy;
 
-final class DebugPass implements CompilerPassInterface
+final class FeatureFlagsDebugPass implements CompilerPassInterface
 {
     public function process(ContainerBuilder $container): void
     {
         if (!$container->has('feature_flags.data_collector')) {
             return;
         }
-
-        $container->register('debug.feature_flags.feature_checker', TraceableFeatureChecker::class)
-            ->setDecoratedService('feature_flags.feature_checker')
-            ->setArguments([
-                '$featureChecker' => new Reference('.inner'),
-                '$dataCollector' => new Reference('feature_flags.data_collector'),
-            ])
-        ;
 
         foreach ($container->findTaggedServiceIds('feature_flags.feature_strategy') as $serviceId => $tags) {
             $container->register('debug.'.$serviceId, TraceableStrategy::class)
