+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Compiler/UnusedTagsPass.php
@@ -48,6 +48,8 @@ class UnusedTagsPass implements CompilerPassInterface
         'controller.targeted_value_resolver',
         'data_collector',
         'event_dispatcher.dispatcher',
+        'feature_flags.feature_provider',
+        'feature_flags.feature_strategy',
         'form.type',
         'form.type_extension',
         'form.type_guesser',
