+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/schema/symfony-1.0.xsd
@@ -45,6 +45,7 @@
             <xsd:element name="enabled-locale" type="xsd:string" minOccurs="0" maxOccurs="unbounded" />
             <xsd:element name="webhook" type="webhook" minOccurs="0" maxOccurs="1" />
             <xsd:element name="remote-event" type="remote-event" minOccurs="0" maxOccurs="1" />
+            <xsd:element name="feature-flags" type="feature-flags" minOccurs="0" maxOccurs="1" />
         </xsd:choice>
 
         <xsd:attribute name="http-method-override" type="xsd:boolean" />
@@ -956,5 +957,25 @@
 
     <xsd:complexType name="remote-event">
         <xsd:attribute name="enabled" type="xsd:boolean" />
+    </xsd:complexType>
+
+    <xsd:complexType name="feature-flags">
+        <xsd:sequence>
+            <xsd:element name="strategies" type="feature-flags_strategy" minOccurs="0" maxOccurs="unbounded" />
+            <xsd:element name="features" type="feature-flags_feature" minOccurs="0" maxOccurs="unbounded" />
+        </xsd:sequence>
+    </xsd:complexType>
+    <xsd:complexType name="feature-flags_strategy">
+        <xsd:sequence>
+            <xsd:element name="with" type="xsd:anyType" minOccurs="0" maxOccurs="unbounded" />
+        </xsd:sequence>
+        <xsd:attribute name="name" type="xsd:string" />
+        <xsd:attribute name="type" type="xsd:string" />
+    </xsd:complexType>
+    <xsd:complexType name="feature-flags_feature">
+        <xsd:attribute name="name" type="xsd:string" />
+        <xsd:attribute name="description" type="xsd:string" />
+        <xsd:attribute name="default" type="xsd:boolean" />
+        <xsd:attribute name="strategy" type="xsd:string" />
     </xsd:complexType>
 </xsd:schema>
