+++ src/Symfony/Component/FeatureFlags/LICENSE
@@ -1,4 +1,4 @@
-Copyright (c) 2004-present Fabien Potencier
+Copyright (c) 2023-present Fabien Potencier
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
