+++ src/Symfony/Component/FeatureFlags/Strategy/HttpFoundation/RequestStackAttributeStrategy.php
@@ -9,24 +9,24 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Bundle\FeatureFlagsBundle\Strategy;
+namespace Symfony\Component\FeatureFlags\Strategy\HttpFoundation;
 
 use Symfony\Component\FeatureFlags\StrategyResult;
 use Symfony\Component\HttpFoundation\Request;
 
-final class RequestStackQueryStrategy extends RequestStackStrategy
+final class RequestStackAttributeStrategy extends RequestStackStrategy
 {
     public function __construct(
-        private readonly string $queryParameterName,
+        private readonly string $attributeName,
     ) {
     }
 
     protected function computeRequest(Request $request): StrategyResult
     {
-        if (false === $request->query->has($this->queryParameterName)) {
+        if (false === $request->attributes->has($this->attributeName)) {
             return StrategyResult::Abstain;
         }
 
-        return $request->query->getBoolean($this->queryParameterName) ? StrategyResult::Grant : StrategyResult::Deny;
+        return $request->attributes->getBoolean($this->attributeName) ? StrategyResult::Grant : StrategyResult::Deny;
     }
 }
