+++ src/Symfony/Component/FeatureFlags/Strategy/HttpFoundation/RequestStackQueryStrategy.php
@@ -9,24 +9,24 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Bundle\FeatureFlagsBundle\Strategy;
+namespace Symfony\Component\FeatureFlags\Strategy\HttpFoundation;
 
 use Symfony\Component\FeatureFlags\StrategyResult;
 use Symfony\Component\HttpFoundation\Request;
 
-final class RequestStackHeaderStrategy extends RequestStackStrategy
+final class RequestStackQueryStrategy extends RequestStackStrategy
 {
     public function __construct(
-        private readonly string $headerName,
+        private readonly string $queryParameterName,
     ) {
     }
 
     protected function computeRequest(Request $request): StrategyResult
     {
-        if (false === $request->headers->has($this->headerName)) {
+        if (false === $request->query->has($this->queryParameterName)) {
             return StrategyResult::Abstain;
         }
 
-        return filter_var($request->headers->get($this->headerName), \FILTER_VALIDATE_BOOL) ? StrategyResult::Grant : StrategyResult::Deny;
+        return $request->query->getBoolean($this->queryParameterName) ? StrategyResult::Grant : StrategyResult::Deny;
     }
 }
